timeseries-ingestion-service/requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
flask==0.12.2
idna==2.6
itsdangerous==0.24
jinja2==2.10
markupsafe==1.0
numpy==1.14.0
pandas==0.22.0
pypugjs==5.1.0
python-dateutil==2.6.1
pytz==2017.3
pyyaml==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
uuid==1.30
websocket-client==0.46.0
werkzeug==0.14.1
